Tema3/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-cloud-datastore==2.19.0
google-cloud-storage==2.16.0
google-cloud-tasks==2.16.0
google-cloud-secret-manager==2.20.0
python-multipart==0.0.9
pydantic==2.6.0